+++ src/components/composite/filter/expert/ExpertFilterForm.tsx
@@ -26,7 +26,7 @@ import {
 import { FieldConstants } from '../../../../utils/constants/fieldConstants';
 import { InputWithPopupConfirmation } from '../../../ui/reactHookForm/selectInputs/InputWithPopupConfirmation';
 import { SelectInput } from '../../../ui/reactHookForm/selectInputs/SelectInput';
-import { CustomReactQueryBuilder } from '../../../ui/reactQueryBuilder/CustomReactQueryBuilder';
+import { CustomReactQueryBuilder } from '../../reactQueryBuilder/CustomReactQueryBuilder';
 import { unscrollableDialogStyles } from '../../../ui/dialogs';
 import { FieldType } from '../../../../utils/types/fieldType';
 import { useFormatLabelWithUnit } from '../../../../hooks/useFormatLabelWithUnit';
