+++ src/components/composite/index.ts
@@ -5,10 +5,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-export * from './customAGGrid';
 export * from './dnd-table';
 export * from './elementSearch';
 export * from './filter';
 export * from './grid';
-export * from './treeViewFinder';
 export * from './muiTable';
+export * from './reactQueryBuilder';
