+++ src/components/composite/reactQueryBuilder/ValueEditor.tsx
@@ -13,14 +13,14 @@ import { useFormContext } from 'react-hook-form';
 import { CountryValueEditor } from './CountryValueEditor';
 import { TranslatedValueEditor } from './TranslatedValueEditor';
 import { TextValueEditor } from './TextValueEditor';
-import { DataType } from '../../composite/filter/expert/expertFilter.type';
+import { DataType } from '../filter/expert/expertFilter.type';
 import { FieldConstants } from '../../../utils/constants/fieldConstants';
 import { Substation, VoltageLevel } from '../../../utils/types/equipmentTypes';
 import { ElementValueEditor } from './ElementValueEditor';
 import { ElementType } from '../../../utils/types/elementType';
 import { PropertyValueEditor } from './PropertyValueEditor';
 import { GroupValueEditor } from './compositeRuleEditor/GroupValueEditor';
-import { OPERATOR_OPTIONS } from '../../composite/filter/expert/expertFilterConstants';
+import { OPERATOR_OPTIONS } from '../filter/expert/expertFilterConstants';
 import { FieldType } from '../../../utils/types/fieldType';
 import { ElementAttributes } from '../../../utils';
 import type { MuiStyles } from '../../../utils/styles';
