+++ src/components/parameters/dynamic-margin-calculation/dynamic-margin-calculation-inline.tsx
@@ -24,8 +24,8 @@ import { SubmitButton } from '../../ui/reactHookForm/utils/SubmitButton';
 import { PopupConfirmationDialog } from '../../ui/dialogs/popupConfirmationDialog/PopupConfirmationDialog';
 import { parametersStyles } from '../parameters-style';
 import { CreateParameterDialog } from '../common';
-import { DirectoryItemSelector } from '../../directoryItemSelector';
-import { TreeViewFinderNodeProps } from '../../composite/treeViewFinder';
+import { DirectoryItemSelector } from '../../ui/directoryItemSelector';
+import { TreeViewFinderNodeProps } from '../../ui/treeViewFinder';
 import { fetchDynamicMarginCalculationParameters } from '../../../services/dynamic-margin-calculation';
 import { useSnackMessage } from '../../../hooks';
 import { CustomFormProvider } from '../../ui';
