+++ src/components/parameters/dynamic-simulation/curve/dialog/equipment-filter.tsx
@@ -15,7 +15,7 @@ import { EquipmentType, VoltageLevelInfos } from '../../../../../utils/types/equ
 import { useLocalizedCountries } from '../../../../../hooks/useLocalizedCountries';
 import { GsLang } from '../../../../../utils/langs';
 import { CheckboxAutocomplete } from '../../../../ui/inputs/checkbox-autocomplete';
-import { CustomAGGrid } from '../../../../composite/customAGGrid';
+import { CustomAGGrid } from '../../../../ui/customAGGrid';
 import { useSnackMessage } from '../../../../../hooks/useSnackMessage';
 import { snackWithFallback } from '../../../../../utils/error';
 import { MuiStyles } from '../../../../../utils/styles';
