+++ src/components/parameters/loadflow/load-flow-parameters-inline.tsx
@@ -13,13 +13,13 @@ import { LoadFlowProvider } from './load-flow-parameters-provider';
 import { parametersStyles } from '../parameters-style';
 import { UseParametersBackendReturnProps } from '../../../utils/types/parameters.type';
 import { ComputingType } from '../common/computing-type';
-import { TreeViewFinderNodeProps } from '../../composite/treeViewFinder';
+import { TreeViewFinderNodeProps } from '../../ui/treeViewFinder';
 import { useSnackMessage } from '../../../hooks';
 import { SubmitButton } from '../../ui';
 import { LabelledButton } from '../common/parameters';
 import { ElementType, GsLang, mergeSx } from '../../../utils';
 import { LineSeparator } from '../common';
-import { DirectoryItemSelector } from '../../directoryItemSelector';
+import { DirectoryItemSelector } from '../../ui/directoryItemSelector';
 import { fetchLoadFlowParameters } from '../../../services/loadflow';
 import { CreateParameterDialog } from '../common/parameters-creation-dialog';
 import { useLoadFlowParametersForm } from './use-load-flow-parameters-form';
