+++ src/components/parameters/pcc-min/pcc-min-parameters-inline.tsx
@@ -16,9 +16,9 @@ import { PopupConfirmationDialog } from '../../ui/dialogs';
 import { UsePccMinParametersForm } from './use-pcc-min-parameters-form';
 import { PccMinParametersForm } from './pcc-min-parameters-form';
 import { fetchPccMinParameters, updatePccMinParameters } from '../../../services/pcc-min';
-import { DirectoryItemSelector } from '../../directoryItemSelector';
+import { DirectoryItemSelector } from '../../ui/directoryItemSelector';
 import { ElementType, mapPccMinParameters, PccMinParametersEnriched, snackWithFallback } from '../../../utils';
-import { TreeViewFinderNodeProps } from '../../composite/treeViewFinder';
+import { TreeViewFinderNodeProps } from '../../ui/treeViewFinder';
 import { fromPccMinParametersFormToParamValuesEnriched, fromPccMinParamsDataToFormValues } from './pcc-min-form-utils';
 
 export function PccMinParametersInLine({
