+++ src/components/parameters/sensi/sensitivity-analysis-parameters-inline.tsx
@@ -20,12 +20,12 @@ import { ComputingType, CreateParameterDialog } from '../common';
 import { useSnackMessage } from '../../../hooks';
 import { SubmitButton } from '../../ui';
 import { parametersStyles } from '../parameters-style';
-import { DirectoryItemSelector } from '../../directoryItemSelector';
+import { DirectoryItemSelector } from '../../ui/directoryItemSelector';
 import {
     fetchSensitivityAnalysisParameters,
     setSensitivityAnalysisParameters,
 } from '../../../services/sensitivity-analysis';
-import { TreeViewFinderNodeProps } from '../../composite/treeViewFinder';
+import { TreeViewFinderNodeProps } from '../../ui/treeViewFinder';
 import { useSensitivityAnalysisParametersForm } from './use-sensitivity-analysis-parameters';
 import { SensitivityAnalysisParametersForm } from './sensitivity-analysis-parameters-form';
 import { PopupConfirmationDialog } from '../../ui/dialogs';
