+++ src/components/ui/dialogs/elementSaveDialog/ElementSaveDialog.tsx
@@ -12,9 +12,9 @@ import { Grid, Box, Button, CircularProgress, Typography } from '@mui/material';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import yup from '../../../../utils/yupConfig';
-import { TreeViewFinderNodeProps } from '../../../composite/treeViewFinder';
+import { TreeViewFinderNodeProps } from '../../treeViewFinder';
 import { DescriptionField, RadioInput, UniqueNameInput } from '../../../ui/reactHookForm';
-import { DirectoryItemSelector } from '../../../directoryItemSelector';
+import { DirectoryItemSelector } from '../../directoryItemSelector';
 import { CustomMuiDialog } from '../customMuiDialog/CustomMuiDialog';
 import { ElementAttributes, ElementType, FieldConstants, MAX_CHAR_DESCRIPTION } from '../../../../utils';
 import { useSnackMessage } from '../../../../hooks';
