+++ src/components/ui/dialogs/modifyElementSelection/ModifyElementSelection.tsx
@@ -11,9 +11,9 @@ import { FormattedMessage, useIntl } from 'react-intl';
 import { useController } from 'react-hook-form';
 import type { UUID } from 'node:crypto';
 import { FolderOutlined } from '@mui/icons-material';
-import { TreeViewFinderNodeProps } from '../../../composite/treeViewFinder';
+import { TreeViewFinderNodeProps } from '../../treeViewFinder';
 import { FieldConstants } from '../../../../utils/constants/fieldConstants';
-import { DirectoryItemSelector } from '../../../directoryItemSelector/DirectoryItemSelector';
+import { DirectoryItemSelector } from '../../directoryItemSelector/DirectoryItemSelector';
 import { ElementType } from '../../../../utils/types/elementType';
 import { fetchDirectoryElementPath } from '../../../../services';
 import { ElementAttributes } from '../../../../utils';
