+++ src/components/ui/reactHookForm/DirectoryItemsInput.tsx
@@ -15,9 +15,9 @@ import { CustomTooltip } from '../../ui/tooltip/CustomTooltip';
 import { useCustomFormContext } from './provider';
 import { ErrorInput, MidFormError } from './errorManagement';
 import { useSnackMessage } from '../../../hooks';
-import { TreeViewFinderNodeProps } from '../../composite/treeViewFinder';
+import { TreeViewFinderNodeProps } from '../treeViewFinder';
 import { type MuiStyles } from '../../../utils/styles';
-import { DirectoryItemSelector } from '../../directoryItemSelector';
+import { DirectoryItemSelector } from '../directoryItemSelector';
 import { fetchDirectoryElementPath } from '../../../services';
 import { ArrayAction, ElementAttributes, getEquipmentTypeShortLabel, mergeSx } from '../../../utils';
 import { NAME } from './constants';
